Refactor getNodePathFromSourceRange
